fix: use rules_rust compile_data rather than data for include_str!
Currently the rust_library data attribute is used to specify source
files that are passed to include_str! macro calls. This results in "No
such file or directory" errors for me when I try to use rules_py via
local_path_override.

rules_rust documentation suggests that the compile_data attribute "can
be used to specify any data files that are embedded into the library,
such as via the include_str! macro" See this link for more info:
https://bazelbuild.github.io/rules_rust/rust.html#rust_binary-compile_data

This PR updates the problematic rust_library to use compile_data instead
of data, which resolves the errors I was running into.